requirements.txt
asyncgui==0.6.3
asynckivy==0.6.4
certifi==2025.1.31
charset-normalizer==3.4.1
docutils==0.21.2
exceptiongroup==1.2.2
filetype==1.2.0
flatbuffers==25.2.10
idna==3.10
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.8.0
Kivy-Garden==0.1.5
kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip
materialyoucolor==2.0.10
numpy==1.26.4
opencv-python==4.11.0.86
pillow==11.1.0
pygments==2.19.1
pypiwin32==223
pywin32==310
requests==2.32.3
tflite==2.18.0
tflite-runtime==2.5.0.post1
urllib3==2.4.0
uv==0.6.14
